Implement tenant name change in dashboard and single services for users/tenants
This commit includes an update in which the hardcoded tenant name in the dashboard has been replaced with a dynamically loaded one. It also introduces services for fetching one user or tenant by their id. In TenantDashboardViewComponent, methods have been added for getting the tenant id from the URL and requesting the associated tenant from the server.