Refactor tenant data fetching and refresh mechanism
Migrated data fetching logic to the TenantTable from TenantView component and introduced a new TableRefresher component. The goal of this shift was to localize the logic surrounding data fetching, and to make the TenantTable component more self-contained and independent. Furthermore, the refresh mechanism now gets triggered through the tenant service to keep the data consistent and current.