Add prettier into project and update related configurations
The "prettier" package was added into the project, and relevant configurations were updated accordingly in package-lock.json and cache related json files. Prettier is an opinionated code formatter that helps to keep code consistent across a team. Adding it will help further maintain the codebase cleaner and easier to read for everyone.