Add search functionality to user table
A search input was added to the User Header component, allowing for user search functionality. The state signal for this search functionality was implemented in the User Service. Other minor changes include the addition of roleId field in the form data of the User Add/Edit dialog component and preparing to filter the user table data based on tenantId.