Add search functionality to customers component
The new functionality allows users to search within the customers component by typing into the search bar. The customer service is integrated with the newly imported BaseSearchComponent and a searchInCustomers method was added to handle search events emitted from the search bar. This search bar is now replacing the previous BaseTableSearch within the customers-header component.