Update Angular and Taiga UI dependencies
Updated the versions of Angular, Taiga UI and other dependencies in package-lock.json file for performance improvement and to take advantage of new features. This includes minor version updates for most dependencies and patches for others, adjusting to the latest stable versions. Dependabot alerts will also be reduced with these updates.