Add Google Maps view to application
This commit adds functionality to view a Google Map in the application. A new Map view is added to the navigation menu, secured with an authentication guard. This also adds map specific components like map header and google map component. It includes styles and translations related to this new feature. The '@angular/google-maps' package was installed and configured for map rendering. Changes also made to adjust vertical spacing in the view layout.