Add date management functionality to components
A date checking pipe was created and integrated into several components to manage different representations of date data. Also, a base date-picker was added and properly set up in applicable components. Customers view and table components were updated to use this new functionality, including the addition of 'Create at' column to the table.