Add clean text field functionality to date-picker
The base date-picker component has been updated to include a ‘clear text field’ functionality. This involves the importation and declaration of TuiTextfieldControllerModule in the component TS file, and setting the 'tuiTextfieldCleaner' to true in the HTML file. Also, additional checks have been added in the customers-header component to handle null dates.