Refactor search methods and introduce base date search
Moved the search logic directly to the base search component, resulting in the removal of individual search methods in user-header, tenant-header, and customers-header components. Furthermore, a base date search component was introduced to add the ability to search based on date. This streamlines the search functionality and provides a more efficient way to implement searching.