Implement conditional rendering for app-navigation and refactor routes
Modified the app-navigation component to render only when the user is logged in. Also, the routes have been refactored to a more efficient structure where the authGuard is provided at the parent level to all protected routes, instead of individually. Also, the AuthService has been injected into the AppComponent to determine the logged-in status of the user.