Add padding mixins, update translations, and enhance components
This commit introduces new padding mixins for SCSS and updates the German and English translation files. It also improves several components, notably the Settings view and its subcomponents. A new enumeration was created for managing languages, and a new SettingsTabsComponent was added. Moreover, the header layout was updated to conditionally display breadcrumbs, and the combobox component was improved with generic types.