Update Angular and Taiga-ui dependencies
The update includes changes in the version of Angular and Taiga-ui dependencies within the package-lock.json file. Angular dependencies have been upgraded from "^18.0.3" to "^18.0.4" and Taiga-ui from "^3.83.0" to "^3.84.0". A few other accompanying updates are also made in devDependencies and integrity checks for these dependencies. This should improve performance or bug fixes that come with the newer versions.