Implement base radio block component
Added a new base radio block component and integrated it into the user-add-edit-dialog component. Updated the styling of header-layout component to include a gap. The base radio block component replaced the previous boolean control to represent user status, vastly improving the usability and readability. The status is now being displayed and manipulated as a string, then converted back to boolean before form submission.