Refactor code to use signals instead of BehaviorSubjects
In this update, we made a comprehensive change in the code to use signals instead of BehaviorSubjects, notably reflecting in tenant, user, and customer table components. This refactor not only simplifies data flow but also leads to improved efficiency. Also, additional instructions were added to the README for using the base-delete-dialog component. This shift effectively modifies the way we handle data flow and sets the ground for more maintainable and understandable code in the long run.