Update Angular and related dependencies to latest versions
This commit upgrades Angular and related dependencies in the `package-lock.json` file to the latest versions, including improvements and bug fixes. Dependencies such as `@angular/animations`, `@angular/core`, and `@angular/cli` have been updated to 19.0.5 or 19.0.6 as appropriate. This ensures the project stays up-to-date with the latest Angular ecosystem changes.